pyproject.toml
[project]
name = "yeppoh"
version = "0.1.0"
description = "RL-driven living sculptures — multi-agent creatures with emergent behavior"
requires-python = ">=3.10,<3.14"
dependencies = [
    "genesis-world",
    "torch>=2.1",
    "torchrl>=0.4",
    "tensordict>=0.4",
    "pettingzoo>=1.24",
    "trimesh>=4.0",
    "numpy>=1.24",
    "scipy>=1.12",
    "omegaconf>=2.3",
    "hydra-core>=1.3",
    "tensorboard>=2.15",
    "matplotlib>=3.8",
    "imageio>=2.33",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
line-length = 100
target-version = "py310"
